Trait domain_runtime_primitives::DomainCoreApi

pub trait DomainCoreApi<Block: BlockT>: Core<Block> {

    // Provided methods
    fn extract_signer(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<<Block as BlockT>::Extrinsic>
    ) -> Result<Vec<(Option<AccountId>, <Block as BlockT>::Extrinsic)>, ApiError> { ... }
    fn is_within_tx_range(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: &<Block as BlockT>::Extrinsic,
        bundle_vrf_hash: &U256,
        tx_range: &U256
    ) -> Result<bool, ApiError> { ... }
    fn intermediate_roots(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Vec<[u8; 32]>, ApiError> { ... }
    fn initialize_block_with_post_state_root(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        header: &<Block as BlockT>::Header
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn apply_extrinsic_with_post_state_root(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: <Block as BlockT>::Extrinsic
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn construct_set_code_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        code: Vec<u8>
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn construct_timestamp_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        moment: Moment
    ) -> Result<Block::Extrinsic, ApiError> { ... }
    fn is_inherent_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: &<Block as BlockT>::Extrinsic
    ) -> Result<bool, ApiError> { ... }
    fn check_transaction_validity(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        uxt: &<Block as BlockT>::Extrinsic,
        block_number: NumberFor<Block>,
        block_hash: <Block as BlockT>::Hash
    ) -> Result<Result<(), CheckTxValidityError>, ApiError> { ... }
    fn storage_keys_for_verifying_transaction_validity(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        account_id: AccountId,
        block_number: NumberFor<Block>,
        tx_era: Option<Era>
    ) -> Result<Result<Vec<Vec<u8>>, VerifyTxValidityError>, ApiError> { ... }
    fn extrinsic_era(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: &<Block as BlockT>::Extrinsic
    ) -> Result<Option<Era>, ApiError> { ... }
    fn extrinsic_weight(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        ext: &Block::Extrinsic
    ) -> Result<Weight, ApiError> { ... }
    fn block_rewards(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Balance, ApiError> { ... }
    fn block_digest(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Digest, ApiError> { ... }
    fn block_weight(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Weight, ApiError> { ... }

}

Base API that every domain runtime must implement.

Provided Methods

fn extract_signer( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<<Block as BlockT>::Extrinsic> ) -> Result<Vec<(Option<AccountId>, <Block as BlockT>::Extrinsic)>, ApiError>

Extracts the optional signer per extrinsic.

fn is_within_tx_range( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: &<Block as BlockT>::Extrinsic, bundle_vrf_hash: &U256, tx_range: &U256 ) -> Result<bool, ApiError>

fn intermediate_roots( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Vec<[u8; 32]>, ApiError>

Returns the intermediate storage roots in an encoded form.

fn initialize_block_with_post_state_root( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, header: &<Block as BlockT>::Header ) -> Result<Vec<u8>, ApiError>

Returns the storage root after initializing the block.

fn apply_extrinsic_with_post_state_root( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: <Block as BlockT>::Extrinsic ) -> Result<Vec<u8>, ApiError>

Returns the storage root after applying the extrinsic.

fn construct_set_code_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, code: Vec<u8> ) -> Result<Vec<u8>, ApiError>

Returns an encoded extrinsic aiming to upgrade the runtime using given code.

fn construct_timestamp_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, moment: Moment ) -> Result<Block::Extrinsic, ApiError>

Returns an encoded extrinsic to set timestamp.

fn is_inherent_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: &<Block as BlockT>::Extrinsic ) -> Result<bool, ApiError>

Returns true if the extrinsic is an inherent extrinsic.

fn check_transaction_validity( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, uxt: &<Block as BlockT>::Extrinsic, block_number: NumberFor<Block>, block_hash: <Block as BlockT>::Hash ) -> Result<Result<(), CheckTxValidityError>, ApiError>

Checks the validity of extrinsic in a bundle.

fn storage_keys_for_verifying_transaction_validity( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, account_id: AccountId, block_number: NumberFor<Block>, tx_era: Option<Era> ) -> Result<Result<Vec<Vec<u8>>, VerifyTxValidityError>, ApiError>

Returns the storage keys of states accessed in the API check_transaction_validity.

fn extrinsic_era( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: &<Block as BlockT>::Extrinsic ) -> Result<Option<Era>, ApiError>

Returns extrinsic Era if present

fn extrinsic_weight( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, ext: &Block::Extrinsic ) -> Result<Weight, ApiError>

Return the extrinsic weight

fn block_rewards( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Balance, ApiError>

The accumulated transaction fee of all transactions included in the block

fn block_digest( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Digest, ApiError>

Return the block digest

fn block_weight( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Weight, ApiError>

Return the consumed weight of the block