Crate domain_runtime_primitives

Common primitives for subspace domain runtime.

Modules

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

AccountId20

A fully Ethereum-compatible AccountId. Conforms to H160 address and ECDSA key standards. Alternative to H256->H160 mapping.

AccountId20Converter

An AccountId20 to MultiAccount converter.

AccountIdConverter

An AccountId32 to MultiAccount converter.

CheckExtrinsicsValidityError

DecodeExtrinsicError

Enums

MultiAccountId

MultiAccountId used by all the domains to describe their account type.

Constants

CHECK_EXTRINSICS_AND_DO_PRE_DISPATCH_METHOD_NAME

fullu qualified method name of check_extrinsics_and_do_pre_dispatch runtime api. Used to call state machine. Change it when the runtime api’s name is changed in the interface.

ERR_BALANCE_OVERFLOW

Custom error when balance overflow occurs.

ERR_NONCE_OVERFLOW

Custom error when nonce overflow occurs.

EXISTENTIAL_DEPOSIT

The existential deposit. Same with the one on primary chain.

NORMAL_DISPATCH_RATIO

Dispatch ratio for domains

SLOT_DURATION

Slot duration that is same as consensus chain runtime.

Traits

Signer

Extracts the signer from an unchecked extrinsic.

TryConvertBack

Extensible conversion trait. Generic over both source and destination types.

Functions

block_weights

calculate_max_bundle_weight

Calculates the max bundle weight

maximum_block_length

Maximum block length for all dispatches. This is set to 3.75 MiB since consensus chain supports on 3.75 MiB for normal

maximum_domain_block_weight

The maximum domain block weight with 3.25 MiB as proof size Consensus allows 3.75 MiB but Fraud proof can carry extra size along with proof size So we set the proof size to 3.25 MiB

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

Balance

Balance of an account.

BlockNumber

An index to a block.

EVMChainId

The EVM chain Id type

Hash

A hash of some data used by the chain.

Nonce

Index of a transaction in the chain.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.