Crate domain_runtime_primitives

Common primitives for subspace domain runtime.

Modules

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Traits

DomainCoreApi

Base API that every domain runtime must implement.

Signer

Extracts the signer from an unchecked extrinsic.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

Address

The address format for describing accounts.

Balance

Balance of an account.

BlockNumber

An index to a block.

Hash

A hash of some data used by the chain.

Index

Index of a transaction in the chain.

RelayerId

The type we use to represent relayer id. This is same as account Id.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.