Trait domain_runtime_primitives::DomainCoreApi

pub trait DomainCoreApi<Block: BlockT, AccountId: Encode + Decode>: Core<Block> {
    // Provided methods
    fn extract_signer(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsics: Vec<<Block as BlockT>::Extrinsic>
    ) -> Result<Vec<(Option<AccountId>, <Block as BlockT>::Extrinsic)>, ApiError> { ... }
    fn extract_signer_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        extrinsics: Vec<<Block as BlockT>::Extrinsic>
    ) -> Result<Vec<(Option<AccountId>, <Block as BlockT>::Extrinsic)>, ApiError> { ... }
    fn intermediate_roots(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash
    ) -> Result<Vec<[u8; 32]>, ApiError> { ... }
    fn intermediate_roots_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext
    ) -> Result<Vec<[u8; 32]>, ApiError> { ... }
    fn initialize_block_with_post_state_root(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        header: &<Block as BlockT>::Header
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn initialize_block_with_post_state_root_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        header: &<Block as BlockT>::Header
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn apply_extrinsic_with_post_state_root(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        extrinsic: <Block as BlockT>::Extrinsic
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn apply_extrinsic_with_post_state_root_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        extrinsic: <Block as BlockT>::Extrinsic
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn construct_set_code_extrinsic(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        code: Vec<u8>
    ) -> Result<Vec<u8>, ApiError> { ... }
    fn construct_set_code_extrinsic_with_context(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        context: ExecutionContext,
        code: Vec<u8>
    ) -> Result<Vec<u8>, ApiError> { ... }
}

Base API that every domain runtime must implement.

Provided Methods

fn extract_signer( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsics: Vec<<Block as BlockT>::Extrinsic> ) -> Result<Vec<(Option<AccountId>, <Block as BlockT>::Extrinsic)>, ApiError>

Extracts the optional signer per extrinsic.

fn extract_signer_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, extrinsics: Vec<<Block as BlockT>::Extrinsic> ) -> Result<Vec<(Option<AccountId>, <Block as BlockT>::Extrinsic)>, ApiError>

Extracts the optional signer per extrinsic.

fn intermediate_roots( &self, __runtime_api_at_param__: <Block as BlockT>::Hash ) -> Result<Vec<[u8; 32]>, ApiError>

Returns the intermediate storage roots in an encoded form.

fn intermediate_roots_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext ) -> Result<Vec<[u8; 32]>, ApiError>

Returns the intermediate storage roots in an encoded form.

fn initialize_block_with_post_state_root( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, header: &<Block as BlockT>::Header ) -> Result<Vec<u8>, ApiError>

Returns the storage root after initializing the block.

fn initialize_block_with_post_state_root_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, header: &<Block as BlockT>::Header ) -> Result<Vec<u8>, ApiError>

Returns the storage root after initializing the block.

fn apply_extrinsic_with_post_state_root( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, extrinsic: <Block as BlockT>::Extrinsic ) -> Result<Vec<u8>, ApiError>

Returns the storage root after applying the extrinsic.

fn apply_extrinsic_with_post_state_root_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, extrinsic: <Block as BlockT>::Extrinsic ) -> Result<Vec<u8>, ApiError>

Returns the storage root after applying the extrinsic.

fn construct_set_code_extrinsic( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, code: Vec<u8> ) -> Result<Vec<u8>, ApiError>

Returns an encoded extrinsic aiming to upgrade the runtime using given code.

fn construct_set_code_extrinsic_with_context( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, context: ExecutionContext, code: Vec<u8> ) -> Result<Vec<u8>, ApiError>

Returns an encoded extrinsic aiming to upgrade the runtime using given code.